examples/github-app-cdk/pyproject.toml
[project]
name = "picofun-github-app-cdk-example"
version = "0.1.0"
description = "PicoFun CDK example using the GitHub API with GitHub App authentication"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib==2.240.0",
    "constructs==10.5.1",
]

[dependency-groups]
dev = [
    "aws-lambda-powertools==3.24.0",
    "boto3==1.42.57",
    "coverage==7.13.4",
    "PyJWT[crypto]==2.11.0",
    "pytest==9.0.2",
    "pytest-mock==3.15.1",
    "ruff==0.15.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["helpers"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
exclude = ["cdk.out"]

[tool.ruff.lint]
select = ["B", "D", "E", "F", "G", "I", "N", "S", "W", "ANN" ,"BLE", "C4", "C90", "DTZ", "ERA", "PLW", "PT", "RET", "RUF", "SIM", "TRY", "UP"]
ignore = ["D203", "D211", "D212", "E501", "F403", "F405"]
fixable = ["B", "D", "E", "F", "G", "I", "N", "S", "W", "ANN" ,"BLE", "C4", "C90", "DTZ", "ERA", "PLW", "PT", "RET", "RUF", "SIM", "TRY", "UP"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S108"]
